moved rate limit functionality into its own file
