+++ exchanges/bitflyer/bitflyer_wrapper.go
@@ -3,7 +3,6 @@ package bitflyer
 import (
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/common"
 	"github.com/thrasher-corp/gocryptotrader/config"
@@ -19,7 +18,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/exchanges/websocket/wshandler"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/withdraw"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
-	"golang.org/x/time/rate"
 )
 
 // GetDefaultConfig returns a default exchange config
@@ -91,54 +89,12 @@ func (b *Bitflyer) SetDefaults() {
 
 	b.Requester = request.New(b.Name,
 		common.NewHTTPClientWithTimeout(exchange.DefaultHTTPTimeout),
-		&RateLimit{
-			Auth: request.NewRateLimit(biflyerRateInterval,
-				bitflyerPrivateRequestRate),
-			UnAuth: request.NewRateLimit(biflyerRateInterval,
-				bitflyerPublicRequestRate),
-			Order: request.NewRateLimit(biflyerRateInterval,
-				bitflyerPrivateSendOrderRequestRate),
-			LowVolume: request.NewRateLimit(time.Minute,
-				bitflyerPrivateLowVolumeRequestRate),
-		})
+		SetRateLimit())
 
 	b.API.Endpoints.URLDefault = japanURL
 	b.API.Endpoints.URL = b.API.Endpoints.URLDefault
 	b.API.Endpoints.URLSecondaryDefault = chainAnalysis
 	b.API.Endpoints.URLSecondary = b.API.Endpoints.URLSecondaryDefault
-}
-
-// RateLimit implements the rate.Limiter interface
-type RateLimit struct {
-	Auth   *rate.Limiter
-	UnAuth *rate.Limiter
-
-	// Send a New Order
-	// Submit New Parent Order (Special order)
-	// Cancel All Orders
-	Order     *rate.Limiter
-	LowVolume *rate.Limiter
-}
-
-// Limit limits outbound requests
-func (r *RateLimit) Limit(f request.EndpointLimit) error {
-	switch f {
-	case request.Auth:
-		time.Sleep(r.Auth.Reserve().Delay())
-	case orders:
-		res := r.Auth.Reserve()
-		time.Sleep(r.Order.Reserve().Delay())
-		time.Sleep(res.Delay())
-	case lowVolume:
-		authShell := r.Auth.Reserve()
-		orderShell := r.Order.Reserve()
-		time.Sleep(r.LowVolume.Reserve().Delay())
-		time.Sleep(orderShell.Delay())
-		time.Sleep(authShell.Delay())
-	default:
-		time.Sleep(r.UnAuth.Reserve().Delay())
-	}
-	return nil
 }
 
 // Setup takes in the supplied exchange configuration details and sets params
