+++ exchanges/bitmex/bitmex_wrapper.go
@@ -5,7 +5,6 @@ import (
 	"math"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/common"
 	"github.com/thrasher-corp/gocryptotrader/config"
@@ -21,7 +20,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/exchanges/websocket/wshandler"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/withdraw"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
-	"golang.org/x/time/rate"
 )
 
 // GetDefaultConfig returns a default exchange config
@@ -137,10 +135,7 @@ func (b *Bitmex) SetDefaults() {
 
 	b.Requester = request.New(b.Name,
 		common.NewHTTPClientWithTimeout(exchange.DefaultHTTPTimeout),
-		&RateLimit{
-			Auth:   request.NewRateLimit(bitmexRateInterval, bitmexAuthRate),
-			UnAuth: request.NewRateLimit(bitmexRateInterval, bitmexUnauthRate),
-		})
+		SetRateLimit())
 
 	b.API.Endpoints.URLDefault = bitmexAPIURL
 	b.API.Endpoints.URL = b.API.Endpoints.URLDefault
@@ -149,22 +144,6 @@ func (b *Bitmex) SetDefaults() {
 	b.WebsocketResponseMaxLimit = exchange.DefaultWebsocketResponseMaxLimit
 	b.WebsocketResponseCheckTimeout = exchange.DefaultWebsocketResponseCheckTimeout
 	b.WebsocketOrderbookBufferLimit = exchange.DefaultWebsocketOrderbookBufferLimit
-}
-
-// RateLimit implements the request.Limiter interface
-type RateLimit struct {
-	Auth   *rate.Limiter
-	UnAuth *rate.Limiter
-}
-
-// Limit limits outbound calls
-func (r *RateLimit) Limit(f request.EndpointLimit) error {
-	if f == request.Auth {
-		time.Sleep(r.Auth.Reserve().Delay())
-		return nil
-	}
-	time.Sleep(r.UnAuth.Reserve().Delay())
-	return nil
 }
 
 // Setup takes in the supplied exchange configuration details and sets params
