+++ exchanges/coinbasepro/coinbasepro.go
@@ -9,7 +9,6 @@ import (
 	"net/url"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/common"
 	"github.com/thrasher-corp/gocryptotrader/common/crypto"
@@ -53,10 +52,6 @@ const (
 	coinbaseproWithdrawalCrypto        = "withdrawals/crypto"
 	coinbaseproCoinbaseAccounts        = "coinbase-accounts"
 	coinbaseproTrailingVolume          = "users/self/trailing-volume"
-
-	coinbaseproRateInterval = time.Second
-	coinbaseproAuthRate     = 5
-	coinbaseproUnauthRate   = 3
 )
 
 // CoinbasePro is the overarching type across the coinbasepro package
