+++ exchanges/coinbene/coinbene.go
@@ -65,83 +65,6 @@ const (
 	coinbeneListSwapPositions  = "/position/list"
 	coinbenePositionFeeRate    = "/position/feeRate"
 
-	// Contract rate limit time interval and request rates
-	contractRateInterval                 = time.Second * 2
-	orderbookContractReqRate             = 20
-	tickersContractReqRate               = 20
-	klineContractReqRate                 = 20
-	tradesContractReqRate                = 20
-	contractAccountInfoContractReqRate   = 10
-	positionInfoContractReqRate          = 10
-	placeOrderContractReqRate            = 20
-	cancelOrderContractReqRate           = 20
-	getOpenOrdersContractReqRate         = 5
-	openOrdersByPageContractReqRate      = 5
-	getOrderInfoContractReqRate          = 10
-	getClosedOrdersContractReqRate       = 5
-	getClosedOrdersbyPageContractReqRate = 5
-	cancelMultipleOrdersContractReqRate  = 5
-	getOrderFillsContractReqRate         = 10
-	getFundingRatesContractReqRate       = 10
-
-	// Spot rate limit time interval and request rates
-	spotRateInterval             = time.Second
-	getPairsSpotReqRate          = 2
-	getPairsInfoSpotReqRate      = 3
-	getOrderbookSpotReqRate      = 6
-	getTickerListSpotReqRate     = 6
-	getSpecificTickerSpotReqRate = 6
-	getMarketTradesSpotReqRate   = 3
-	// getKlineSpotReqRate              = 1
-	// getExchangeRateSpotReqRate       = 1
-	getAccountInfoSpotReqRate        = 3
-	queryAccountAssetInfoSpotReqRate = 6
-	placeOrderSpotReqRate            = 6
-	batchOrderSpotReqRate            = 3
-	queryOpenOrdersSpotReqRate       = 3
-	queryClosedOrdersSpotReqRate     = 3
-	querySpecficOrderSpotReqRate     = 6
-	queryTradeFillsSpotReqRate       = 3
-	cancelOrderSpotReqRate           = 6
-	cancelOrdersBatchSpotReqRate     = 3
-
-	// Rate limit functionality
-	contractOrderbook request.EndpointLimit = iota
-	contractTickers
-	contractKline
-	contractTrades
-	contractAccountInfo
-	contractPositionInfo
-	contractPlaceOrder
-	contractCancelOrder
-	contractGetOpenOrders
-	contractOpenOrdersByPage
-	contractGetOrderInfo
-	contractGetClosedOrders
-	contractGetClosedOrdersbyPage
-	contractCancelMultipleOrders
-	contractGetOrderFills
-	contractGetFundingRates
-
-	spotPairs
-	spotPairInfo
-	spotOrderbook
-	spotTickerList
-	spotSpecificTicker
-	spotMarketTrades
-	spotKline        // Not implemented yet
-	spotExchangeRate // Not implemented yet
-	spotAccountInfo
-	spotAccountAssetInfo
-	spotPlaceOrder
-	spotBatchOrder
-	spotQueryOpenOrders
-	spotQueryClosedOrders
-	spotQuerySpecficOrder
-	spotQueryTradeFills
-	spotCancelOrder
-	spotCancelOrdersBatch
-
 	limitOrder    = "1"
 	marketOrder   = "2"
 	buyDirection  = "1"
