+++ exchanges/hitbtc/hitbtc_wrapper.go
@@ -6,7 +6,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/thrasher-corp/gocryptotrader/common"
 	"github.com/thrasher-corp/gocryptotrader/config"
@@ -22,7 +21,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/exchanges/websocket/wshandler"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/withdraw"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
-	"golang.org/x/time/rate"
 )
 
 // GetDefaultConfig returns a default exchange config
@@ -115,11 +113,7 @@ func (h *HitBTC) SetDefaults() {
 
 	h.Requester = request.New(h.Name,
 		common.NewHTTPClientWithTimeout(exchange.DefaultHTTPTimeout),
-		&RateLimit{
-			MarketData: request.NewRateLimit(hitbtcRateInterval, hitbtcMarketDataReqRate),
-			Trading:    request.NewRateLimit(hitbtcRateInterval, hitbtcTradingReqRate),
-			Other:      request.NewRateLimit(hitbtcRateInterval, hitbtcAllOthers),
-		})
+		SetRateLimit())
 
 	h.API.Endpoints.URLDefault = apiURL
 	h.API.Endpoints.URL = h.API.Endpoints.URLDefault
@@ -128,28 +122,6 @@ func (h *HitBTC) SetDefaults() {
 	h.WebsocketResponseMaxLimit = exchange.DefaultWebsocketResponseMaxLimit
 	h.WebsocketResponseCheckTimeout = exchange.DefaultWebsocketResponseCheckTimeout
 	h.WebsocketOrderbookBufferLimit = exchange.DefaultWebsocketOrderbookBufferLimit
-}
-
-// RateLimit implements the request.Limiter interface
-type RateLimit struct {
-	MarketData *rate.Limiter
-	Trading    *rate.Limiter
-	Other      *rate.Limiter
-}
-
-// Limit limits outbound requests
-func (r *RateLimit) Limit(f request.EndpointLimit) error {
-	switch f {
-	case marketRequests:
-		time.Sleep(r.MarketData.Reserve().Delay())
-	case tradingRequests:
-		time.Sleep(r.Trading.Reserve().Delay())
-	case otherRequests:
-		time.Sleep(r.Other.Reserve().Delay())
-	default:
-		return errors.New("functionality not found")
-	}
-	return nil
 }
 
 // Setup sets user exchange configuration settings
