+++ exchanges/poloniex/poloniex.go
@@ -49,10 +49,6 @@ const (
 	poloniexLendingHistory       = "returnLendingHistory"
 	poloniexAutoRenew            = "toggleAutoRenew"
 
-	poloniexRateInterval = time.Second
-	poloniexAuthRate     = 6
-	poloniexUnauthRate   = 6
-
 	poloniexDateLayout = "2006-01-02 15:04:05"
 )
