+++ exchanges/poloniex/poloniex_wrapper.go
@@ -21,7 +21,6 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/exchanges/websocket/wshandler"
 	"github.com/thrasher-corp/gocryptotrader/exchanges/withdraw"
 	log "github.com/thrasher-corp/gocryptotrader/logger"
-	"golang.org/x/time/rate"
 )
 
 // GetDefaultConfig returns a default exchange config
@@ -113,14 +112,7 @@ func (p *Poloniex) SetDefaults() {
 
 	p.Requester = request.New(p.Name,
 		common.NewHTTPClientWithTimeout(exchange.DefaultHTTPTimeout),
-		// If your account's volume is over $5 million in 30 day volume,
-		// you may be eligible for an API rate limit increase.
-		// Please email poloniex@circle.com.
-		// As per https://docs.poloniex.com/#http-api
-		&RateLimit{
-			Auth:   request.NewRateLimit(poloniexRateInterval, poloniexAuthRate),
-			UnAuth: request.NewRateLimit(poloniexRateInterval, poloniexUnauthRate),
-		})
+		SetRateLimit())
 
 	p.API.Endpoints.URLDefault = poloniexAPIURL
 	p.API.Endpoints.URL = p.API.Endpoints.URLDefault
@@ -129,22 +121,6 @@ func (p *Poloniex) SetDefaults() {
 	p.WebsocketResponseMaxLimit = exchange.DefaultWebsocketResponseMaxLimit
 	p.WebsocketResponseCheckTimeout = exchange.DefaultWebsocketResponseCheckTimeout
 	p.WebsocketOrderbookBufferLimit = exchange.DefaultWebsocketOrderbookBufferLimit
-}
-
-// RateLimit implements the request.Limiter interface
-type RateLimit struct {
-	Auth   *rate.Limiter
-	UnAuth *rate.Limiter
-}
-
-// Limit limits outbound calls
-func (r *RateLimit) Limit(f request.EndpointLimit) error {
-	if f == request.Auth {
-		time.Sleep(r.Auth.Reserve().Delay())
-		return nil
-	}
-	time.Sleep(r.UnAuth.Reserve().Delay())
-	return nil
 }
 
 // Setup sets user exchange configuration settings
